# docs/openapi/agent.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "jpcite Agent Evidence API",
    "description": "Agent-safe OpenAPI subset for evidence prefetch before answer generation. It excludes billing, webhook, OAuth, account-management, and operator endpoints. Anonymous callers can evaluate within the published daily limit unless an operation marks X-API-Key as required; callers that need higher volume send X-API-Key.",
    "version": "0.3.2",
    "contact": {
      "name": "jpcite Support",
      "url": "https://jpcite.com/tokushoho.html"
    },
    "termsOfService": "https://jpcite.com/tos.html",
    "license": {
      "name": "Proprietary - see termsOfService"
    }
  },
  "paths": {
...
    "/v1/evidence/packets/query": {
      "post": {
        "tags": [
          "evidence"
        ],
        "summary": "Evidence Packet — multi-record query composer",
        "description": "Use this endpoint as source-linked evidence prefetch before GPT, Claude, Cursor, or RAG answer generation. It returns a compact Evidence Packet instead of a final narrative answer, so callers can avoid pasting long PDFs, official pages, or search snippets into the model. 1 packet = 1 billable unit; see the pricing page for current public price and anonymous limits. The packet bundles up to `limit` records (hard cap 500). Truncation surfaces `_warning=\"truncated\"`. Optional compression fields are input-context estimates, not external provider billing guarantees.",
        "operationId": "post_evidence_packet_query_v1_evidence_packets_query_post",
        "parameters": [
          {
            "name": "output_format",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "json",
                "csv",
                "md"
              ],
              "type": "string",
              "description": "`json` (default) / `csv` / `md`.",
              "default": "json",
              "title": "Output Format"
            },
            "description": "`json` (default) / `csv` / `md`."
          },
          {
            "name": "X-API-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EvidencePacketQueryBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvidencePacketEnvelope"
                },
                "example": {
                  "packet_id": "evp_example",
                  "generated_at": "2026-05-02T12:00:00+09:00",
                  "api_version": "v1",
                  "corpus_snapshot_id": "snap_20260502",
                  "query": {
                    "user_intent": "Tokyo manufacturer subsidy evidence",
                    "normalized_filters": {
                      "prefecture": "Tokyo"
                    }
                  },
                  "answer_not_included": true,
                  "records": [
                    {
                      "entity_id": "program:example",
                      "primary_name": "Example public program",
                      "record_kind": "program",
                      "source_url": "https://example.go.jp/program",
                      "source_fetched_at": "2026-05-01T00:00:00+09:00",
                      "source_checksum": "sha256:example",
                      "license": "gov_standard_v2.0",
                      "authority_name": "Example authority",
                      "precomputed": {
                        "basis": "am_program_summary",
                        "summaries": {
                          "200": "Short source-linked evidence summary."
                        }
                      }
                    }
                  ],
                  "quality": {
                    "freshness_bucket": "current",
                    "coverage_score": 0.86,
                    "known_gaps": [],
                    "human_review_required": false
                  },
                  "verification": {
                    "replay_endpoint": "/v1/programs/search?q=...",
                    "freshness_endpoint": "/v1/meta/freshness"
                  },
                  "compression": {
                    "packet_tokens_estimate": 566,
                    "source_tokens_estimate": 14000,
                    "avoided_tokens_estimate": 13434,
                    "compression_ratio": 0.0404,
                    "input_context_reduction_rate": 0.9596,
                    "source_tokens_basis": "pdf_pages",
                    "source_pdf_pages": 20,
                    "estimate_scope": "input_context_only",
                    "savings_claim": "estimate_not_guarantee",
                    "provider_billing_not_guaranteed": true,
                    "cost_savings_estimate": {
                      "currency": "JPY",
                      "input_token_price_jpy_per_1m": 300,
                      "gross_input_savings_jpy": 4.0,
                      "jpcite_billable_units": 1,
                      "jpcite_cost_jpy_ex_tax": 3,
                      "net_savings_jpy_ex_tax": 1.0,
                      "break_even_avoided_tokens": 10000,
                      "break_even_source_tokens_estimate": 10566,
                      "break_even_met": true,
                      "input_context_only": true,
                      "price_input_source": "caller_supplied",
                      "billing_savings_claim": "estimate_not_guarantee",
                      "provider_billing_not_guaranteed": true
                    }
                  },
                  "evidence_value": {
                    "records_returned": 1,
                    "source_linked_records": 1,
                    "precomputed_records": 1,
                    "pdf_fact_refs": 0,
                    "known_gap_count": 0,
                    "fact_provenance_coverage_pct_avg": 0.86,
                    "web_search_performed_by_jpcite": false,
                    "request_time_llm_call_performed": false
                  }
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          },
          {}
        ],
        "x-jpcite-agent-safe": true,
        "x-jpcite-auth": "optional_x_api_key_for_paid_volume",
        "x-jpcite-agent-priority": 2,
        "x-jpcite-route-purpose": "source_linked_evidence_packet"
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "EvidencePacketQueryBody": {
        "properties": {
          "query_text": {
            "type": "string",
            "maxLength": 500,
            "minLength": 1,
            "title": "Query Text",
            "description": "Free-text query. Echoed into `query.user_intent`."
          },
          "filters": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Filters",
            "description": "Optional structured filters (prefecture / tier). Echoed into `query.normalized_filters`."
          },
          "limit": {
            "type": "integer",
            "maximum": 500.0,
            "minimum": 1.0,
            "title": "Limit",
            "description": "Cap on records[] length. Hard cap = 500.",
            "default": 10
          },
          "include_facts": {
            "type": "boolean",
            "title": "Include Facts",
            "description": "Include records[].facts[] source-linked fact records.",
            "default": true
          },
          "include_rules": {
            "type": "boolean",
            "title": "Include Rules",
            "description": "Include records[].rules[] compatibility/exclusion rules.",
            "default": false
          },
          "include_compression": {
            "type": "boolean",
            "title": "Include Compression",
            "description": "Include input-context size estimates. Estimates compare the packet against a caller-supplied source baseline; they are not provider billing guarantees.",
            "default": false
          },
          "fields": {
            "type": "string",
            "title": "Fields",
            "description": "Field projection level. `default` / `full`.",
            "default": "default"
          },
          "input_token_price_jpy_per_1m": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Input Token Price Jpy Per 1M",
            "description": "Optional caller-supplied input-token price in JPY per 1M tokens. Used only for an input-context break-even reference; not a total provider bill estimate."
          },
          "source_tokens_basis": {
            "type": "string",
            "enum": [
              "unknown",
              "pdf_pages",
              "token_count"
            ],
            "title": "Source Tokens Basis",
            "description": "Caller-supplied baseline for context comparison. `unknown` returns packet size only. `pdf_pages` uses source_pdf_pages * 700 tokens/page as an estimate. `token_count` uses source_token_count exactly as supplied by the caller.",
            "default": "unknown"
          },
          "source_pdf_pages": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 1000.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Pdf Pages",
            "description": "PDF page count the caller would otherwise paste/fetch into the LLM. Used only when source_tokens_basis=pdf_pages."
          },
          "source_token_count": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 50000000.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Token Count",
            "description": "Caller-measured token count for the source context the LLM would otherwise read. Used only when source_tokens_basis=token_count."
          }
        },
        "type": "object",
        "required": [
          "query_text"
        ],
        "title": "EvidencePacketQueryBody"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "Customer API key issued via Stripe Checkout. Anonymous tier (no key) gets 3 req/日 per IP."
      }
    }
  },
  "servers": [
    {
      "url": "https://api.jpcite.com",
      "description": "Production"
    }
  ],
  "security": []
}