# src/jpintel_mcp/api/intelligence.py
"""Precomputed intelligence REST surface.

This router is intentionally thin: it exposes the compact, precomputed
Evidence Packet path used by the offline token-cost benchmark without
adding an LLM dependency or a second datastore.
"""

from __future__ import annotations

import time
from typing import Annotated, Any, Literal

from fastapi import APIRouter, Query
from fastapi.responses import JSONResponse, Response

from jpintel_mcp.api._audit_seal import attach_seal_to_body
from jpintel_mcp.api._response_models import (
    PRECOMPUTED_INTELLIGENCE_EXAMPLE,
    PrecomputedIntelligenceBundle,
)
from jpintel_mcp.api.deps import ApiContextDep, DbDep, log_usage
from jpintel_mcp.api.evidence import (
    _gate_evidence_envelope,
    _get_composer,
    _validate_compression_baseline,
)
from jpintel_mcp.services.evidence_packet import (
    MAX_RECORDS_PER_PACKET,
    EvidencePacketComposer,
)


def _build_evidence_value(envelope: dict[str, Any], records: list[Any]) -> dict[str, Any]:
    """Plan §4-A `evidence_value` thin wrapper around the composer's helper.

    Centralises the call so the precomputed bundle, the regular Evidence
    Packet, and any future surface all return the same shape.
    """
    safe_records = [rec for rec in records if isinstance(rec, dict)]
    return EvidencePacketComposer._build_evidence_value(envelope, safe_records)

router = APIRouter(prefix="/v1/intelligence", tags=["intelligence"])


def _annotate_precomputed_bundle(envelope: dict[str, Any]) -> dict[str, Any]:
    """Add benchmark-facing metadata to an Evidence Packet envelope."""
    records = envelope.get("records") or []
    precomputed_count = sum(1 for rec in records if rec.get("precomputed"))
    gaps = envelope.setdefault("quality", {}).setdefault("known_gaps", [])
    if (
        precomputed_count == 0
        and isinstance(gaps, list)
        and "precomputed_summary_unavailable" not in gaps
    ):
        gaps.append("precomputed_summary_unavailable")

    packet_id = str(envelope.get("packet_id") or "")
    bundle_id = packet_id.replace("evp_", "pci_", 1) if packet_id else ""
    envelope.update(
        {
            "bundle_kind": "precomputed_intelligence",
            "bundle_id": bundle_id,
            "answer_basis": ("precomputed" if precomputed_count else "metadata_only"),
            "records_returned": len(records),
            "precomputed_record_count": precomputed_count,
            "precomputed": {
                "available": precomputed_count > 0,
                "basis_tables": ["am_program_summary"],
                "record_count": precomputed_count,
                "note": (
                    "Precomputed summaries are deterministic corpus artifacts. "
                    "They are not generated by a request-time LLM."
                ),
            },
            "usage": {
                "intended_for": "llm_context_prefetch",
                "web_search_required": False,
                "jpcite_requests": 1,
                "billing_units": 1,
            },
        }
    )
    envelope["agent_recommendation"] = _build_agent_recommendation(
        records_returned=len(records),
        precomputed_count=precomputed_count,
        compression=envelope.get("compression"),
        envelope=envelope,
    )
    # Plan §4-A: surface AI-readable evidence_value counters even on the
    # precomputed bundle path. The composer already attaches one; we
    # recompute here so caller-visible counters stay in lock-step with
    # the envelope's bundle_kind / records_returned.
    envelope["evidence_value"] = _build_evidence_value(envelope, records)
    return envelope


def _source_linked_count(envelope: dict[str, Any] | None) -> int:
    """Count records that carry a source_url or a source-linked fact."""
    if not isinstance(envelope, dict):
        return 0
    records = envelope.get("records") or []
    count = 0
    for rec in records:
        if not isinstance(rec, dict):
            continue
        if rec.get("source_url"):
            count += 1
            continue
        facts = rec.get("facts") or []
        if any(isinstance(f, dict) and f.get("source") for f in facts):
            count += 1
    return count


def _build_agent_recommendation(
    *,
    records_returned: int,
    precomputed_count: int,
    compression: Any,
    envelope: dict[str, Any] | None = None,
) -> dict[str, Any]:
    """Return machine-readable guidance for LLM/tool agents.

    This does not claim guaranteed provider-cost savings. It makes the
    already-present evidence/compression signals easy for an agent to branch on.

    Plan §4-A two-axis decision:
        * recommend_for_evidence / evidence_decision (baseline-free)
        * recommend_for_cost_savings / cost_savings_decision (caller baseline)
    """
    context_savings = _context_savings_summary(compression)
    source_linked = _source_linked_count(envelope)
    if records_returned <= 0:
        evidence_decision = "no_records_returned"
        recommend_for_evidence = False
    elif source_linked <= 0:
        evidence_decision = "records_returned_without_source_links"
        recommend_for_evidence = False
    else:
        evidence_decision = "supported_by_source_linked_records"
        recommend_for_evidence = True

    quality = (envelope or {}).get("quality") if isinstance(envelope, dict) else None
    gap_list = (
        quality.get("known_gaps")
        if isinstance(quality, dict) and isinstance(quality.get("known_gaps"), list)
        else []
    )

    if records_returned <= 0:
        return {
            "recommendation": "broaden_query_or_skip",
            "recommend_to_user": False,
            "recommend_for_evidence": False,
            "evidence_decision": evidence_decision,
            "route": "/v1/intelligence/precomputed/query",
            "reason_codes": [
                "no_records_returned",
                "query_needs_broader_terms_or_different_endpoint",
            ],
            "value_reasons": [],
            "message": (
                "Do not recommend jpcite for this exact query yet; broaden "
                "the query or use a more specific endpoint."
            ),
            "context_savings": context_savings,
            **_cost_savings_decision(context_savings),
        }

    if not recommend_for_evidence:
        return {
            "recommendation": "broaden_query_or_skip",
            "recommend_to_user": False,
            "recommend_for_evidence": False,
            "evidence_decision": evidence_decision,
            "route": "/v1/intelligence/precomputed/query",
            "reason_codes": [
                "records_returned_without_source_links",
                "query_needs_broader_terms_or_different_endpoint",
            ],
            "value_reasons": [],
            "message": (
                "Do not recommend jpcite for this exact query yet; the "
                "bundle lacks source-linked evidence."
            ),
            "context_savings": context_savings,
            **_cost_savings_decision(context_savings),
        }

    if precomputed_count > 0:
        reason_codes = [
            "precomputed_summary_available",
            "source_linked_records_returned",
            "no_request_time_llm",
            "no_live_web_search",
        ]
        if context_savings and context_savings.get("break_even_met") is True:
            reason_codes.append("caller_baseline_break_even_met")
        elif context_savings and context_savings.get("evaluated") is True:
            reason_codes.append("caller_baseline_evaluated")
        else:
            reason_codes.append("context_savings_baseline_needed")

        value_reasons = ["source_linked_records_returned", "precomputed_summary_available"]
        if gap_list:
            value_reasons.append("known_gaps_exposed")
        value_reasons.extend(["no_request_time_llm", "no_live_web_search"])
        if context_savings and context_savings.get("break_even_met") is True:
            value_reasons.append("caller_baseline_break_even_met")

        return {
            "recommendation": "use_jpcite_prefetch",
            "recommend_to_user": True,
            "recommend_for_evidence": recommend_for_evidence,
            "evidence_decision": evidence_decision,
            "route": "/v1/intelligence/precomputed/query",
            "reason_codes": reason_codes,
            "value_reasons": value_reasons,
            "message": (
                "Use jpcite before sending long Japanese public-program "
                "sources to an LLM; this bundle provides compact, "
                "source-linked context without a request-time LLM call."
            ),
            "context_savings": context_savings,
            **_cost_savings_decision(context_savings),
        }

    value_reasons = ["source_linked_records_returned"] if recommend_for_evidence else []
    if gap_list:
        value_reasons.append("known_gaps_exposed")
    value_reasons.extend(["no_request_time_llm", "no_live_web_search"])
    if context_savings and context_savings.get("break_even_met") is True:
        value_reasons.append("caller_baseline_break_even_met")

    return {
        "recommendation": "use_evidence_packet",
        "recommend_to_user": True,
        "recommend_for_evidence": recommend_for_evidence,
        "evidence_decision": evidence_decision,
        "route": "/v1/evidence/packets/query",
        "reason_codes": [
            "source_linked_records_returned",
            "precomputed_summary_unavailable",
            "evidence_packet_available",
        ],
        "value_reasons": value_reasons,
        "message": (
            "Recommend jpcite for source-linked evidence prefetch, but treat context "
            "savings as unproven until a comparable baseline is supplied."
        ),
        "context_savings": context_savings,
        **_cost_savings_decision(context_savings),
    }


def _context_savings_summary(compression: Any) -> dict[str, Any] | None:
    if not isinstance(compression, dict):
        return None
    savings = compression.get("cost_savings_estimate")
    evaluated = compression.get("source_tokens_estimate") is not None
    summary: dict[str, Any] = {
        "evaluated": evaluated,
        "source_tokens_basis": compression.get("source_tokens_basis"),
        "source_tokens_input_source": compression.get("source_tokens_input_source"),
        "packet_tokens_estimate": compression.get("packet_tokens_estimate"),
        "source_tokens_estimate": compression.get("source_tokens_estimate"),
        "avoided_tokens_estimate": compression.get("avoided_tokens_estimate"),
        "compression_ratio": compression.get("compression_ratio"),
        "input_context_reduction_rate": compression.get("input_context_reduction_rate"),
        "provider_billing_not_guaranteed": compression.get(
            "provider_billing_not_guaranteed", True
        ),
        "savings_claim": compression.get("savings_claim"),
    }
    if isinstance(savings, dict):
        summary.update(
            {
                "input_token_price_jpy_per_1m": savings.get("input_token_price_jpy_per_1m"),
                "break_even_avoided_tokens": savings.get("break_even_avoided_tokens"),
                "break_even_source_tokens_estimate": savings.get(
                    "break_even_source_tokens_estimate"
                ),
                "break_even_met": savings.get("break_even_met"),
                "net_savings_jpy_ex_tax": savings.get("net_savings_jpy_ex_tax"),
                "billing_savings_claim": savings.get("billing_savings_claim"),
            }
        )
    return summary


def _cost_savings_decision(context_savings: dict[str, Any] | None) -> dict[str, Any]:
    """Tell agents whether a cost-savings recommendation is justified."""
    if not context_savings or context_savings.get("evaluated") is not True:
        return {
            "recommend_for_cost_savings": False,
            "cost_savings_decision": "needs_caller_baseline",
            "missing_for_cost_claim": [
                "source_tokens_basis",
                "source_pdf_pages_or_source_token_count",
                "input_token_price_jpy_per_1m",
            ],
        }

    if context_savings.get("break_even_met") is True:
        return {
            "recommend_for_cost_savings": True,
            "cost_savings_decision": "supported_by_caller_baseline",
            "missing_for_cost_claim": [],
        }

    if context_savings.get("break_even_met") is False:
        return {
            "recommend_for_cost_savings": False,
            "cost_savings_decision": "not_supported_by_caller_baseline",
            "missing_for_cost_claim": [],
        }

    return {
        "recommend_for_cost_savings": False,
        "cost_savings_decision": "needs_input_token_price",
        "missing_for_cost_claim": ["input_token_price_jpy_per_1m"],
    }


@router.get(
    "/precomputed/query",
    summary="Precomputed Intelligence Bundle — query prefetch",
    description=(
        "Returns a compact precomputed intelligence bundle for LLM context "
        "prefetch. Use when an agent needs short source-linked Japanese "
        "public-program evidence before answering, without live web search "
        "or a request-time LLM call. Optional compression fields are "
        "input-context estimates, not external provider billing guarantees."
    ),
    responses={
        200: {
            "model": PrecomputedIntelligenceBundle,
            "content": {"application/json": {"example": PRECOMPUTED_INTELLIGENCE_EXAMPLE}},
        }
    },
)
def get_precomputed_intelligence_query(
    q: Annotated[
        str,
        Query(
            min_length=1,
            max_length=500,
            description="Free-text query to resolve against precomputed program summaries.",
        ),
    ],
    conn: DbDep,
    ctx: ApiContextDep,
    prefecture: Annotated[
        str | None,
        Query(description="Optional prefecture filter."),
    ] = None,
    tier: Annotated[
        str | None,
        Query(description="Optional tier filter."),
    ] = None,
    limit: Annotated[
        int,
        Query(
            ge=1,
            le=MAX_RECORDS_PER_PACKET,
            description=(f"Cap on records[] length. Hard cap = {MAX_RECORDS_PER_PACKET}."),
        ),
    ] = 10,
    include_facts: Annotated[
        bool,
        Query(
            description=(
                "Include raw facts in addition to precomputed summaries. "
                "Default false keeps the bundle compact for LLM context."
            ),
        ),
    ] = False,
    include_compression: Annotated[
        bool,
        Query(description="Include deterministic token-compression estimate."),
    ] = True,
    input_token_price_jpy_per_1m: Annotated[
        float | None,
        Query(
            description=(
                "Optional caller input-token price in JPY per 1M tokens. "
                "Used only for an optional reference comparison; no savings "
                "or cost reduction is guaranteed."
            ),
        ),
    ] = None,
    source_tokens_basis: Annotated[
        Literal["unknown", "pdf_pages", "token_count"],
        Query(
            description=(
                "Optional caller-supplied baseline for context comparison. "
                "`unknown` returns packet size only. `pdf_pages` uses "
                "source_pdf_pages * 700 tokens/page as an estimate. "
                "`token_count` uses source_token_count exactly as supplied. "
                "This is input-context estimation only, not a guarantee."
            ),
        ),
    ] = "unknown",
    source_pdf_pages: Annotated[
        int | None,
        Query(
            ge=1,
            le=1000,
            description=(
                "PDF page count the caller would otherwise paste/fetch into "
                "the LLM. Used only when source_tokens_basis=pdf_pages."
            ),
        ),
    ] = None,
    source_token_count: Annotated[
        int | None,
        Query(
            ge=1,
            le=50_000_000,
            description=(
                "Caller-measured token count for the source context the LLM "
                "would otherwise read. Used only when "
                "source_tokens_basis=token_count."
            ),
        ),
    ] = None,
    output_format: Annotated[
        Literal["json"],
        Query(description="Only `json` is supported for benchmark bundles."),
    ] = "json",
) -> Response:
    _t0 = time.perf_counter()
    _validate_compression_baseline(
        source_tokens_basis,
        source_pdf_pages,
        source_token_count,
    )
    filters: dict[str, Any] = {}
    if prefecture:
        filters["prefecture"] = prefecture
    if tier:
        filters["tier"] = tier

    composer = _get_composer()
    envelope = composer.compose_for_query(
        q,
        filters,
        limit=limit,
        include_facts=include_facts,
        include_rules=False,
        include_compression=include_compression,
        fields="default",
        input_token_price_jpy_per_1m=input_token_price_jpy_per_1m,
        source_tokens_basis=source_tokens_basis,
        source_pdf_pages=source_pdf_pages,
        source_token_count=source_token_count,
    )
    envelope, _gate_summary = _gate_evidence_envelope(envelope)
    envelope = _annotate_precomputed_bundle(envelope)

    latency_ms = int((time.perf_counter() - _t0) * 1000)
    log_usage(
        conn,
        ctx,
        "intelligence.precomputed.query",
        latency_ms=latency_ms,
        result_count=len(envelope.get("records") or []),
        params={
            "q": q,
            "limit": limit,
            "filter_keys": sorted(filters),
            "source_tokens_basis": source_tokens_basis,
            "source_pdf_pages": source_pdf_pages,
            "source_token_count": source_token_count,
        },
    )
    attach_seal_to_body(
        envelope,
        endpoint="intelligence.precomputed.query",
        request_params={
            "q": q,
            "limit": limit,
            "filter_keys": sorted(filters),
            "source_tokens_basis": source_tokens_basis,
            "source_pdf_pages": source_pdf_pages,
            "source_token_count": source_token_count,
        },
        api_key_hash=ctx.key_hash,
        conn=conn,
    )
    return JSONResponse(content=envelope)


__all__ = ["router"]